requirements.txt
celery==5.3.4
redis==5.0.1
reportlab==4.4.10
flower==2.0.1
python-dotenv==1.0.0